fix: improve firebase test account deletion process

+++ packages/near-fast-auth-signer-e2e-tests/utils/firebase.ts
@@ -39,7 +39,7 @@ export const deleteAccount = async (userUid: string) => {
 export const deleteUserByEmail = async (email: string) => {
   try {
     const user = await admin.auth().getUserByEmail(email);
-    if (user) {
+    if (user && user.uid) {
       await deleteAccount(user.uid);
     }
   } catch (error) {
